+++ front/src/App/OpenSpace/DisplayTalks.js
@@ -1,23 +1,30 @@
-import { usePushToNewTalk } from '#helpers/routes';
+import { RedirectToRoot, usePushToNewTalk } from '#helpers/routes';
 import EmptyTalk from '../MyTalks/EmptyTalk';
 import { TrackWithTalks } from './TrackWithTalks';
 import TalksGrid from './TalksGrid';
-import React from 'react';
+import React, { useCallback } from 'react';
+import { useGetTalks } from '#api/os-client';
+import Spinner from '#shared/Spinner';
 
 export function DisplayTalks({ amountOfTalks, activeCallForPapers, tracks }) {
+  const { data: talks, isPending, isRejected, reload: reloadTalks } = useGetTalks();
+  const reload = useCallback(() => {
+    reloadTalks();
+  }, [reloadTalks]);
   const pushToNewTalk = usePushToNewTalk();
   const shouldDisplayEmptyTalk = amountOfTalks === 0 && activeCallForPapers;
   const shouldDisplayTrackWithTalks = tracks.length > 0 && amountOfTalks > 0;
-
+  if (isPending) return <Spinner />;
+  if (isRejected) return <RedirectToRoot />;
   if (shouldDisplayEmptyTalk) {
     return <EmptyTalk onClick={pushToNewTalk} />;
   }
 
   if (shouldDisplayTrackWithTalks) {
     return tracks.map((track) => (
-      <TrackWithTalks track={track} activeCallForPapers={activeCallForPapers} />
+      <TrackWithTalks talks={talks} reloadTalks={reload} track={track} />
     ));
   }
 
-  return <TalksGrid />;
+  return <TalksGrid talks={talks} reloadTalks={reload} />;
 }
